requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.30.0
python-dotenv>=1.0.0
python-docx>=1.1.0
google-cloud-bigquery>=3.20.0
google-auth>=2.27.0
db-dtypes>=1.2.0
